Add shardingsphere-db-discovery-distsql-handler module
